// projects/quality-review/src/lib/oper-stat/model/rpdma99.model.ts
import { MapLine } from './map-line.model';

/**
 * Model class Rpdma99
 * Path: screenbean/qltyrvwoperstat
 * Model: com::uhc::aarp::fox::domain::screenbean::qltyrvwoperstat::Rpdma99
 */
export class Rpdma99 {
  m99cmnd = '';
  m99mem1 = '';
  m99mem2 = '';
  m99mem3 = '';
  m99cno1 = '';
  m99cno2 = '';
  m99cno3 = '';
  m99cno4 = '';
  m99cno5 = '';
  m99cno6 = '';
  m99com1 = '';
  m99com2 = '';
  m99bdat = '';
  m99edat = '';
  m99lsit = '';
  m99site = '';
  m99ldiv = '';
  m99div = '';
  m99lloc = '';
  m99loc = '';
  m99last = '';
  m99ast1 = '';
  m99ast2 = '';
  m99ast3 = '';
  m99ast4 = '';
  m99ast5 = '';
  m99ast6 = '';
  m99ast7 = '';
  m99lpst = '';
  m99pst1 = '';
  m99pst2 = '';
  m99pst3 = '';
  m99pst4 = '';
  m99pst5 = '';
  m99pst6 = '';
  m99pst7 = '';
  m99na01 = '';
  m99oi01 = '';
  m99ol01 = '';
  m99oh01 = '';
  m99av01 = '';
  m99rs01 = '';
  m99rv01 = '';
  m99ts01 = '';
  m99ps01 = '';
  m99rm01 = '';
  m99ch01 = '';
  m99qe01 = '';
  m99na02 = '';
  m99oi02 = '';
  m99ol02 = '';
  m99oh02 = '';
  m99av02 = '';
  m99rs02 = '';
  m99rv02 = '';
  m99ts02 = '';
  m99ps02 = '';
  m99rm02 = '';
  m99ch02 = '';
  m99qe02 = '';
  m99na03 = '';
  m99oi03 = '';
  m99ol03 = '';
  m99oh03 = '';
  m99av03 = '';
  m99rs03 = '';
  m99rv03 = '';
  m99ts03 = '';
  m99ps03 = '';
  m99rm03 = '';
  m99ch03 = '';
  m99qe03 = '';
  m99na04 = '';
  m99oi04 = '';
  m99ol04 = '';
  m99oh04 = '';
  m99av04 = '';
  m99rs04 = '';
  m99rv04 = '';
  m99ts04 = '';
  m99ps04 = '';
  m99rm04 = '';
  m99ch04 = '';
  m99qe04 = '';
  m99na05 = '';
  m99oi05 = '';
  m99ol05 = '';
  m99oh05 = '';
  m99av05 = '';
  m99rs05 = '';
  m99rv05 = '';
  m99ts05 = '';
  m99ps05 = '';
  m99rm05 = '';
  m99ch05 = '';
  m99qe05 = '';
  m99na06 = '';
  m99oi06 = '';
  m99ol06 = '';
  m99oh06 = '';
  m99av06 = '';
  m99rs06 = '';
  m99rv06 = '';
  m99ts06 = '';
  m99ps06 = '';
  m99rm06 = '';
  m99ch06 = '';
  m99qe06 = '';
  m99na07 = '';
  m99oi07 = '';
  m99ol07 = '';
  m99oh07 = '';
  m99av07 = '';
  m99rs07 = '';
  m99rv07 = '';
  m99ts07 = '';
  m99ps07 = '';
  m99rm07 = '';
  m99ch07 = '';
  m99qe07 = '';
  m99na08 = '';
  m99oi08 = '';
  m99ol08 = '';
  m99oh08 = '';
  m99av08 = '';
  m99rs08 = '';
  m99rv08 = '';
  m99ts08 = '';
  m99ps08 = '';
  m99rm08 = '';
  m99ch08 = '';
  m99qe08 = '';
  m99na09 = '';
  m99oi09 = '';
  m99ol09 = '';
  m99oh09 = '';
  m99av09 = '';
  m99rs09 = '';
  m99rv09 = '';
  m99ts09 = '';
  m99ps09 = '';
  m99rm09 = '';
  m99ch09 = '';
  m99qe09 = '';
  m99na10 = '';
  m99oi10 = '';
  m99ol10 = '';
  m99oh10 = '';
  m99av10 = '';
  m99rs10 = '';
  m99rv10 = '';
  m99ts10 = '';
  m99ps10 = '';
  m99rm10 = '';
  m99ch10 = '';
  m99qe10 = '';
  m99na11 = '';
  m99oi11 = '';
  m99ol11 = '';
  m99oh11 = '';
  m99av11 = '';
  m99rs11 = '';
  m99rv11 = '';
  m99ts11 = '';
  m99ps11 = '';
  m99rm11 = '';
  m99ch11 = '';
  m99qe11 = '';
  m99thr1 = '';
  m99tav1 = '';
  m99trs1 = '';
  m99trv1 = '';
  m99tsp1 = '';
  m99tps1 = '';
  m99tnr1 = '';
  m99tec1 = '';
  m99tqe1 = '';
  m99err1 = '';
  m99pfk1 = '';
  m99pfk2 = '';
  mapLines: MapLine[] = [];
}