// projects/quality-review/src/lib/oper-stat/model/map-line.model.ts
/**
 * Model class MapLine
 * Path: screenbean/qltyrvwoperstat
 * Model: com::uhc::aarp::fox::domain::screenbean::qltyrvwoperstat::MapLine
 * Legacy Mapping: MAP-LINE
 */
export class MapLine {
  name = '';
  ions = '';
  salLevel = '';
  totalHrs = '';
  avgClmHr = '';
  remPerHhs = '';
  clmsRvwd = '';
  clmsSusp = '';
  percentageSusp = '';
  errRMvd = '';
  errChgd = '';
  qltyExpd = '';
  slno: number | string = 0;
}